moving ExceptionInfo to models
