+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core2/__init__.py
@@ -1,8 +1,8 @@
 from ._action import Action
 from ._marker import mark_step
+from ._models import ExceptionInfo
 from ._workflow import Workflow
 from ._workflow_context import WorkflowContext
-from ._workflow_runner import ExceptionInfo
 from ._workflow_runner_manager import WorkflowRunnerManager
 
 __all__: tuple[str, ...] = (
