+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core2/_models.py
@@ -1,3 +1,12 @@
+from pydantic import BaseModel
+
 StepName = str
 ActionName = str
 WorkflowName = str
+
+
+class ExceptionInfo(BaseModel):
+    exception_class: type
+    action_name: ActionName
+    step_name: StepName
+    serialized_traceback: str
